+++ CHANGELOG.md
@@ -42,7 +42,9 @@ All `0.1.0-alpha` releases are unstable, and have missing widgets.
   - `/unblacklist <login>` was renamed to `/blacklist remove <login>`
   - Added `/blacklist clear` which removes all players from the blacklist.
   - `/map_import <id/uid>` was renamed to `/import map <id/uid>`
-  
+  - Added `/kick <login/nick>` which kicks a player from the server.
+  - Added `/bounce <login/nick>` which moves a player to spectator mode.
+
 ### [Commits](https://github.com/timwie/steward/compare/v0.1.0-alpha5...v0.1.0-alpha6)
 
 <br>